* Separate the custom BSDR UI class from the BlockedShutdownUX interface class * Make the minimum dialog height accurate * Make the HiDPI button drawing accurate * Add a workaround for LogonUI not exiting on cancel when the BSDR phase is forced on the default desktop with WH mod
